Use a static pool size and don't try to grow it dynamically.
Forget about the timeouts - is there an app in the pool or not?
The pool is set to a fixed size, and grows to that size on boot, serving requests as they come in with whatever's in the pool.
Doesn't attempt to spin up new instances to handle requests as they come in.  Also, no timeouts - if there's not an app in the pool available to serve the request, immediately return a 503 for HAProxy to try the request somewhere else - don't wait around for an app to become available.